app/requirements.txt
flask==2.3.2
mysql-connector-python==8.1.0
